OPHJOD-636: Update Tools structure to have common "results"/opportunities
